+++ src/App.jsx
@@ -1,12 +1,18 @@
+import { Routes, Route } from 'react-router-dom'
 import HomePage from './components/HomePage'
+import BlogPostPage from './pages/BlogPostPage'
+import { CursorLogPage } from './pages/CursorLogPage'
 
 function App() {
   return (
-    <div className="center ph3 ph4-ns">
-      <HomePage />
+    <div>
+      <Routes>
+        <Route path="/" element={<HomePage />} />
+        <Route path="/blog/:slug" element={<BlogPostPage />} />
+        <Route path="/logs/:logSlug" element={<CursorLogPage />} />
+      </Routes>
     </div>
   )
 }
 
 export default App
-